moved opening times functions etc. to /common
