+++ common/package.json
@@ -15,6 +15,7 @@
     "fontfaceobserver": "^2.0.13",
     "immutable": "^3.8.2",
     "lazysizes": "^4.0.1",
+    "lodash.groupby": "^4.6.0",
     "moment": "^2.20.1",
     "moment-timezone": "^0.5.14",
     "next": "^5.0.0",
