+++ common/utils/format-date.js
@@ -19,3 +19,10 @@ export function formatDate(date: Date): string {
 export function formatTime(date: Date): string {
   return london(date).format('HH:mm');
 }
+
+export function isDatePast(date: Date): boolean {
+  const momentNow = london();
+  const momentEnd = london(date);
+
+  return momentEnd.isBefore(momentNow);
+}
